HACK: Use fsconfig() and lowerdir+ fd's for overlayfs if possible
This is a quick hack patch to see how much better this performs, in my
testing so far its a lot better.

With user namespaces enabled overlayfs does a lot of mounting currently.
The prior approach was something like:

    1. bind mount / idmap each layer of the overlayfs
    2. Construct the overlayfs with a mount() call
    3. umount() all the bind mounts from 1

That's a mount + umount per layer per container. I've seen situations
where that's 100 pods, 55 layers, so 5500 mounts and umounts. That
really pollutes the mount namespace containerd runs in. For me, systemd
is there too and that hurts systemd as well (since it processes the
mount table and has to compete for access in that period, etc). All
these mounts also effect containerd, slowing the whole thing down to the
point where I've seen the default CRI timeout get hit by kubelet.

Instead let's prototype using the "new" mount api. Let's hack and slash
it all into the shim here, extending upon the already extraordinary
condition of overlayfs + uidmap. We'll now:

    1. open_tree() + mount_setattr() to get an idmap'ed fd for each
       layer of the overlayfs
    2. use those fd's with fsconfig()'s lowerdir+
    3. Shove the rest of the options in with fsconfig()
    4. move_mount() it onto the host for runc to use, etc

The advantage here is that:

    1. We don't pollute the mount namespace
    2. We don't have to take nearly as many locks in do_lock_mount()
       on the kernel side

so this ends up scaling much better.